requirements.txt
Flask==2.0.1
PyJWT
requests
